refactor: split provider settings transport
